+++ src/cobraflex/launch/lane_keeper_gazebo.launch.py
@@ -122,7 +122,7 @@ def generate_launch_description():
             DeclareLaunchArgument(
                 "world",
                 default_value=PathJoinSubstitution(
-                    [FindPackageShare("cobraflex"), "worlds", "lane_following_oval.world"]
+                    [FindPackageShare("cobraflex"), "worlds", "lane_following_complex_b.world"]
                 ),
                 description="World file to load in Gazebo.",
             ),
@@ -133,12 +133,12 @@ def generate_launch_description():
             ),
             DeclareLaunchArgument(
                 "spawn_x",
-                default_value="0",
+                default_value="-2.8", #oval: 0.0 a: -3.111 b: -2.8
                 description="Robot spawn X position.",
             ),
             DeclareLaunchArgument(
                 "spawn_y",
-                default_value="-0.1225",
+                default_value="-1.5", #oval: -0.1225 a: 1.105 b: -1.5
                 description="Robot spawn Y position.",
             ),
             DeclareLaunchArgument(
@@ -148,7 +148,7 @@ def generate_launch_description():
             ),
             DeclareLaunchArgument(
                 "spawn_yaw",
-                default_value="0",
+                default_value="-0.321", #oval: 0.0 a: 0.755 b: -0.321
                 description="Robot spawn yaw in radians.",
             ),
             gazebo_launch,